aws/vector-database/application/requirements.txt
opensearch-py==2.3.2
requests-aws4auth==1.2.0
psycopg2-binary==2.9.7
boto3==1.34.0
pydantic==2.5.0 